try fix (#371)
