+++ apps/container/vite.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'vite';
 import { fileURLToPath } from 'node:url';
 
 export default defineConfig({
+  base: './',
   server: {
     port: 3000,
     host: true,
